+++ application/controllers/admin/users/wholesale/Deactivate.php
@@ -38,12 +38,27 @@ class Deactivate extends Admin_Controller {
 		$this->load->library('users/wholesale_user_details');
 		$this->wholesale_user_details->initialize(array('user_id'=>$id));
 
-		// remove user from mailgun list
-		// only basix for now
-		$params['address'] = $this->wholesale_user_details->email;
-		$params['list_name'] = 'wholesale_users@mg.shop7thavenue.com';
-		$this->load->library('mailgun/list_member_delete', $params);
-		$res = $this->list_member_delete->delete();
+		// get list name for mailgun udpate on all status
+		switch ($this->wholesale_user_details->reference_designer)
+		{
+			case 'tempoparis':
+				$list_name = 'ws_tempo@mg.shop7thavenue.com';
+			break;
+			case 'basixblacklabel':
+				$list_name = 'wholesale_users@mg.shop7thavenue.com';
+			break;
+			default:
+				$list_name = '';
+		}
+
+		if ($list_name)
+		{
+			// remove user from mailgun list
+			$params['address'] = $this->wholesale_user_details->email;
+			$params['list_name'] = $list_name;
+			$this->load->library('mailgun/list_member_delete', $params);
+			$res = $this->list_member_delete->delete();
+		}
 
 		// udpate record
 		$DB = $this->load->database('instyle', TRUE);
